requirements.txt
streamlit
python-dotenv
PyPDF2
langchain-text-splitters
langchain-huggingface
langchain-community
langchain-groq
langchain>=0.2.0
chromadb==0.5.3
protobuf==4.25.3
sentence-transformers